ai_pristup/requirements.txt
torch>=2.1
transformers>=4.44
accelerate>=0.33
huggingface_hub>=0.23
Pillow>=10.2
python-dateutil>=2.9.0
pdf2image>=1.17
jsonschema>=4.22
